+++ app/src/main/java/com/itsabugnotafeature/fitocrazy/ui/home/workout/addExercise/EnterTextForNewExerciseFragment.kt
@@ -16,7 +16,7 @@ import androidx.fragment.app.setFragmentResult
 import com.itsabugnotafeature.fitocrazy.R
 
 
-class EnterTextForNewExerciseFragment(private val hintForExerciseComponentType: String) : DialogFragment() {
+class EnterTextForNewExerciseFragment(private val hintForExerciseComponentType: String, private val existingComponentText: String? = null) : DialogFragment() {
 
     private var userInputtedString: String? = null
 
@@ -33,6 +33,8 @@ class EnterTextForNewExerciseFragment(private val hintForExerciseComponentType: 
 
         view.findViewById<TextView>(R.id.label_enterNewExerciseComponentText).text = getString(R.string.label_enterNewExerciseComponentText, hintForExerciseComponentType.lowercase())
         val editText = view.findViewById<EditText>(R.id.textEntry_newExerciseComponentText)
+
+        if (!existingComponentText.isNullOrBlank()) editText.setText(existingComponentText)
 
         editText.hint = hintForExerciseComponentType
         editText.requestFocus()